Module rrr_core::chart

Structs

Beat

Stores all of the note rows that represent a beat.

BinChart

A space/memory efficient representation of a chart.

ColorIter

An iterator over the variants of Self

Note

Represents a single note in a note row.

NoteRow

Stores the notes that belong to a single row in a beat.

RuntimeChart

A runtime efficient representation of a chart used by an RRR instance.

RuntimeNote

SwfParser

Enums

NoteColor

NoteDirection